backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
httpx>=0.27
pydantic>=2.6
pydantic-settings>=2.2
python-dotenv==1.0.0
sqlmodel>=0.0.14
sqlalchemy>=2.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-httpx==0.21.3
